Remove problematic javascript compiler
This compiler has proven to be useless to us. It showed inconsistencies
and the original developer did not fix them after a request.

If someone would like to show a fully tested Javascript-based compiler,
I'll be happy to merge it. But at this time I feel uncomfortable with
these compilers and the huge, largely incompetent community who can
barely write 4 lines of code.

Signed-off-by: MegaXLR <admin@megaxlr.net>