Reverted changes and removed rawgit.com
The reason i removed those domains that i was having problems with downloading attachments from gmail. and that the domain google used could not be resolved. this resulted that i was a moron and removed malware domains. 

This also closes #22

And thanks to the people that mentioned in commit: 7f1e3c60e88478ff79949833ae3bebccbd15d7b3
I goofed up.